requirements.txt
anyio==4.6.2.post1
appdirs==1.4.4
asttokens==2.4.1
beautifulsoup4==4.12.3
certifi==2024.8.30
colorama==0.4.6
comm==0.2.2
cssselect==1.2.0
debugpy==1.8.8
decorator==5.1.1
et_xmlfile==2.0.0
executing==2.1.0
fake-useragent==1.5.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-html==0.11.0.dev0
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.3.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parse==1.20.2
parso==0.8.4
platformdirs==4.3.6
playwright==1.48.0
prompt_toolkit==3.0.48
psutil==6.1.0
pure_eval==0.2.3
pyee==12.0.0
Pygments==2.18.0
pyppeteer==2.0.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
pyzmq==26.2.0
selectolax==0.3.26
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
stack-data==0.6.3
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
w3lib==2.2.1
wcwidth==0.2.13
websockets==10.4
zipp==3.21.0
